requirements.txt
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.3
Flask==3.0.2
tweepy==4.14.0
python-dotenv==1.0.1
